+++ card/lib/card/content/diff/l_c_s/processor.rb
