+++ card/lib/card/query.rb
@@ -36,7 +36,6 @@ class Card
   module Query
     # # require_dependency "card/query/card_query"
     Card::Query::CardQuery
-    binding.pry
     # After conversion, ATTRIBUTES is a Hash where the key is the attribute
     # and the value is the attribute type:
     # { id: :basic, name: :basic, key: :basic ...}
