+++ card/lib/card/set/advanced_api.rb
@@ -1,3 +1,5 @@
+require_dependency 'card/dirty'
+
 class Card
   module Set
     # advanced set module API
